replaced poetry with uv and restructured project
