--- poetry.toml
@@ -1,4 +1,0 @@
-[virtualenvs]
-create = true
-virtualenvs.in-project = true
-
