+++ pyproject.toml
@@ -17,47 +17,77 @@ classifiers = [
     "Programming Language :: Python",
 ]
 
-[tool.poetry]
-package-mode = false
-
-
-[tool.poetry.dependencies]
-python = ">=3.12,<3.13"
-logly = ">=0.0.6,<0.1.0"
-packaging = ">=24.0,<25.0"
-pandas = "^2.2"
-numpy = "^1.26"
-torch = {version = "^2.5.1", source = "pytorch-gpu"}
-torchvision = {version = "^0.20.1", source = "pytorch-gpu"}
-torchaudio = {version = "^2.5.1", source = "pytorch-gpu"}
-requests = "^2.32"
-tqdm = "^4.67"
-regex = "^2024.11"
-sentencepiece = "^0.2"
-tiktoken = "^0.8"
-blobfile = "^3.0"
-safetensors = "^0.5"
-huggingface-hub = "^0.27"
-xformers = [
-    {  url = "https://download.pytorch.org/whl/cu124/xformers-0.0.29.post1-cp312-cp312-win_amd64.whl"}
+dependencies = [
+  "torch>=2.5.1",
+  "torchvision>=0.20.1",
+  "torchaudio>=2.5.1",
+  "logly>=0.0.6,<0.1.0",
+  "packaging>=24.0,<25.0",
+  "pandas>=2.2",
+  "numpy>=1.26",
+  "requests>=2.32",
+  "tqdm>=4.67",
+  "regex>=2024.11",
+  "sentencepiece>=0.2",
+  "tiktoken>=0.8",
+  "blobfile<=3.0",
+  "safetensors>=0.5",
+  "huggingface-hub>=0.27",
+  "gradio>=5.12",
+  "transformers>=4.48",
+  "datasets>=3.2",
+  "trl>=0.13",
+  "peft>=0.14",
+  "bitsandbytes>=0.45",
+  "unsloth-zoo>=2025.1.3",
+  "accelerate>=1.2.1",
+  "gradio-client>=1.5.4",
+  "toml>=0.10.2",
+  "unsloth[huggingface,cu124onlytorch251]>=2025.1.5",
+  "xformers>=0.0.29.post1",
+  "triton>=3.1.0",
 ]
-gradio = ">=5.12"
-transformers = "^4.48"
-datasets = "^3.2"
-trl = "^0.13"
-peft = "^0.14"
-bitsandbytes = "^0.45"
-unsloth = { extras = ["cu124onlytorch251", "huggingface"], version = "^2025.1.6" }
-unsloth-zoo = " ^2025.1.3"
-accelerate = "^1.2.1"
-gradio-client = ">=1.5.4"
-triton = {url = "https://github.com/woct0rdho/triton-windows/releases/download/v3.1.0-windows.post8/triton-3.1.0-cp312-cp312-win_amd64.whl"}
-toml = "^0.10.2"
 
-[[tool.poetry.source]]
-name = "pytorch-gpu"
+
+
+
+[tool.uv.sources]
+torch = [
+  { index = "pytorch-cu124", marker = "platform_system == 'Windows'" },
+  { index = "pytorch-cu124", marker = "platform_system == 'Linux'" },
+]
+torchvision = [
+  { index = "pytorch-cu124", marker = "platform_system == 'Windows'" },
+  { index = "pytorch-cu124", marker = "platform_system == 'Linux'" },
+]
+torchaudio = [
+  { index = "pytorch-cu124", marker = "platform_system == 'Windows'" },
+    { index = "pytorch-cu124", marker = "platform_system == 'Linux'" },
+    ]
+triton = [
+  {  url="https://github.com/woct0rdho/triton-windows/releases/download/v3.1.0-windows.post8/triton-3.1.0-cp312-cp312-win_amd64.whl", marker = "platform_system == 'Windows'"},
+]
+xformers=[
+    { index = "xformers", marker = "platform_system == 'Windows'" },
+    { index = "xformers", marker = "platform_system == 'Linux'" },
+]
+
+[[tool.uv.index]]
+name = "xformers"
 url = "https://download.pytorch.org/whl/cu124"
-priority = "explicit"
+explicit = true
+
+
+
+[[tool.uv.index]]
+name = "pytorch-cpu"
+url = "https://download.pytorch.org/whl/cpu"
+explicit = true
+
+[[tool.uv.index]]
+name = "pytorch-cu124"
+url = "https://download.pytorch.org/whl/cu124"
+explicit = true
 
 [project.urls]
 source = "https://github.com/muhammad-fiaz/finetune-web-ui.git"
@@ -65,14 +95,21 @@ homepage = "https://github.com/muhammad-fiaz/finetune-web-ui#readme"
 documentation = "https://github.com/muhammad-fiaz/finetune-web-ui#readme"
 repository = "https://github.com/muhammad-fiaz/finetune-web-ui"
 
-[tool.poetry.group.dev.dependencies]
-jupyter = "^1.1.1"
-ipykernel = "^6.29.5"
-poetry-plugin-export = "^1.9.0"
-poetry-plugin-bundle= "^1.5.0"
-poetry-plugin-shell= "^1.0.1"
 
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[project.optional-dependencies]
+dev = [
+    "jupyter>=1.1.1",
+    "ipykernel>=6.29.5",
+]
+
+[tool.uv.pip]
+extra = ["dev"]
+python = "3.12"
+output-file = "requirements.txt"
+
+
+
+[tool.uv]
+environments = ["sys_platform == 'windows'", "sys_platform == 'linux'"]
+cache-keys = [{ file = "requirements.txt" }]
