+++ ui/components/Export.py
@@ -1,0 +1,12 @@
+import gradio as gr
+
+
+class Export:
+    def __init__(self):
+        self.block = None
+        self.options = None
+
+    def create_ui(self):
+        """Create the export UI."""
+        with gr.Group() as export_block:
+            pass
